backend/requirements.txt
fastapi
uvicorn[standard]
pillow
numpy
opencv-python-headless
scikit-image
scikit-learn
torch
torchvision
facenet-pytorch
python-multipart
psutil
reportlab
mahotas
pywavelets
opencv-python
redis>=4.6.0
pytest-asyncio
transformers>=4.0.0